# api/swagger/swagger.yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: Notification Engine API
host: 'localhost:10010'
basePath: /ne
schemes:
  - http
  - https
consumes:
  - multipart/form-data
  - application/json;version=1
produces:
  - application/json
  - text/plain
paths:
...
  '/changeStream/{id}':
    x-swagger-router-controller: controller
    put:
      description: callback response of webHook
      operationId: changeStream
      parameters:
        - name: id
          in: path
          type: string
          required: true
          description: Id of the webHook to be displayed
      responses:
        '200':
          description: List of the entites 
        '500':
          description: Internal server error
...